refactor(test): テストファイルを src/tests/ へ一元集約
各モジュールに隣接していた *_tests.rs / tests.rs を src/tests/ 配下へ集約し、
「テストがどこにあるか」を一目で分かるようにした。

Cargo の `tests/`（パッケージ直下）は結合テスト＝公開APIのみで private に
触れられないため使えない。本コミットは単体テストのまま `#[path]` で各モジュール
配下の tests 子モジュールとして取り込む方式にしており、private アクセスと
`super::*` 参照はそのまま維持される。挙動・バイナリは不変、テストも全件緑。

- src/tests/{config,logger,dashboard,router,watcher,placeholder,win_runas}.rs
- src/tests/actions_{command,common,copy,execute,move}.rs
- 共通ヘルパ test_support は crate 直下の `mod test_support`（src/tests/support.rs）へ移し、
  参照を crate::test_support に統一。
