+++ cat-watcher/src/tests/actions_copy.rs
@@ -1,5 +1,5 @@
 use super::*;
-use crate::actions::test_support::{base_action, make_retry, make_sink, write_file};
+use crate::test_support::{base_action, make_retry, make_sink, write_file};
 use crate::config::ActionType;
 use tempfile::tempdir;
