+++ cat-watcher/src/tests/dashboard.rs
