+++ cat-watcher/src/tests/logger.rs
