+++ cat-watcher/src/tests/watcher.rs
